rename ConvertTo->FromInternal, ConvertFrom->ToInternal, reads better
